Successfully handling data between storage and firebase
This update successfully allows user to generate their recipe plan and save the data to Firebase and local storage.

A property has been added to each Day localStorage which states whether a day 'isUpToDate'. This value becomes false after a recipe is removed and checkStorage() is called to see if there any days in localStorage are no longer up-to-date.

setIsUpToDate is never called when a recipe is added as it should be added to Firebase first, then inside the same method, a call to storage service will add it to localStorage.

isUpToDate will be set to false for a day when the user has removed a recipe from their plan. A call to Firebase will remove the recipe from the specified day which will trigger the Firebase listener for the given day to be triggered to call checkStorage(). checkStorage() will see that the isUpToDate value is false and will adjust localStorage to match with what's in Firebase.