Edamam service added, alterations to checkStorage()
Timetable service will now use a method called searchForRecipe() from the Edamam service instead of when it was all packed into the generate() method. This has been quite a seamless process and everything seems to be working as intended.

This update also includes alterations made to the checkStorage() method in Timetable:

There is now a method added in auth service which is called checkDayMatchesBetweenStorageAndFirebase(uid, dayName).

This method checks if there's any anomalies between Firebase and the localStorage. If there is, it sets the isUpToDate value to false for the day in localStorage.

This method will be run everytime checkStorage is called (I believe every time the timetable page is refreshed) and if any anomalies are spotted, these will be corrected in checkStorage() when it sees isUpToDate is false for the current day it's checking.