Further improvements to SearchForRecipes features
SearchForRecipes feature included, close to being finished.

Material Angular added to project, can now use spinners whilst waiting for results to return

getRemainingNutirents in timetable service now takes a day index and returns the day's remaining nutrients instead of looping through each day and returning everyday's nutrients

addRecipesFromFirebase in storage service will now try to complete the request to Edamam and will catch any errors. This is needed because Edamam seems to delete recipes from their database so if we search for a recipe which no longer exists, it will crash the 'generate' method and prevent us adding recipes appropriately.

A searchForRecipes method has now been included in the Edamam service which instead of returning just one result like the searchForRecipe method. It'll return a given amount of results in the form of a double array. [i] represents the page in the array and [i][j] represents the result in the page.

Finally, a temporary adjustment has been made to the main html in order to test the Whisk Widget feature. Simply just a HTML page formatted to suit their 'Content Validator' along with a script feature which should show a button which when clicked will take the user to a shopping list.