setRemainingNutrients update + form validation, noOfRecipes partially removed
setRemainingNutrients initally took the entire remaining values for a day and set them to the valuess of the textboxes however now we pass the day to the getRecipePlanStructure() method in order to find out how many meals the user has left to find and divide each total remaining nutrient to itself divided by the amount of meals left.

noOfRecipes attribute in each day has partially been removed, it's no longer checked for in the getRecipePlanStructure (instead we just use the length of the recipes array). We no longer increment/decrement when a recipe is added or removed howeever the attribute still exists in the day object. Should probably be removed.