Fixed bug with navigating from home to meal page
Initially, the timetable would only load the carousel and progress bars dynamically if the timetable route was refreshed.

Navigating from home to timetable wouldn't work because the component wasn't alive to receive the emitted event to tell it to load the progress and  carousel.

The code to load the progress and carousel has now been moved to it's own method and is still called from within the subscription method however now loaded from the ngAfterView so the progress and carousel gets loaded regardless of whether the subscribe method receives an event.