Few big updates
-Redundant componets removed

-Recipe component added
--Recipe component retrieves parameters for day and recipeID from route parameters and uses them to find the recipeDetails from the timetable Services.

Storage service now retrieves ingredients when it searches for recipe from Edamam

Timetable Service now has a dataLoaded variable which is assigned to true after the checkStorage() method is called. This allows the Recipe component to check whether or not to retrieve the recipeDetails depending on if the data is available to be called.