Few changes - Instructions, ingredient divider, slight change to recipe html
Instructions component was added in an attempt to show the recipe page in an iFrame however the cors-work-around doesn't seem to be working with angular therefore this may be redundant.

A service which has functions to divide the ingredients has been added and is in progress.

Recipe component's html now includes two small cols which help in centering the main content.