Attempt to fix ingredients not showing in Github pages #5
This should be the final fix, the problem seems to have occured because the user's localStorage was out-of-date and therefore there wasn't an existing ingredients array in the recipe object.

This update will clear the localStorage when the timetable page is opened which should reset the broken recipes.

This is error is important to remember whenever the localStorage structure is changed as it can potentially happen again in the future.