Attempt to fix ingredients not showing in Github pages #6
Hopefully final tweak, just clearing the localStorage wasn't enough because there's a checkDayIsUpToDate variable in timetable.service.checkStorage() which would prevent the site from updating localStorage on my phone.

I'll have to look into the engineering behind the storage system to see if there's a way of ensuring this value is properly working - it might be the case that because it's still seeing the same recipe IDs it considers them to match between firebase and localStorage.

After this update is performed, I'll have to comment out the localStorage.clear() method from timetable service constructor.

And remove the comments from checkStorage() in timetable service for the checkDayIsUpToDate logic.

Perhaps the best way moving forward would to try clearing cache on my phone instead of going through the hassle of commenting in/out these pieces of code when an update is performed on the structure within localStorage - especially when it's the recipe structure.