Ingredient Divider service complete (not perfect), progress component renamed, progress compenent added
Storage and Edamam services now call methods from ingredient divider before adding the ingredients to the recipe in storage. The methods still aren't perfect and it seems like it's going to be nearly impossible to give appropriate ingredient fractions everytime. It'd probably be a good idea to allow the user to flip between the original servings and servings for one person.

Progress was originally used for the homepage however this has now been renamed to welcome to allow progress to be used to display graphs based on the user's progress.