Shopping list component added - close to finished, Bug fix in timetableweeklypage, searchForRecipe() method bug fixed in Edamam service
Shopping list component

Added and display almost finished. User can check which ingredients which they want to include.

To finalise - 1. need to add all selected ingredients into a separate array which can then be pushed into the Whisk array.

2. Need to make it so when user checks the day checkbox, it automatically selects all the recipes & ingredients for that day. Also when the user selects just the recipe, to check the day checkbox and all the ingredients belonging to that recipe.

Timetableweeklypage bug fix

The navigateToRecipe method wasn't sending the correct parameters to the router in order for the recipe component to extract the details from the allRecipes data.

searchForRecipe() in Edamam

This method is called by the generate() method in timetable service. Initially it wasn't updated to parse/convert the ingredients but now it does.