Adjust addRecipeToStorage & addRecipesFromFirebase, added getRecipeByEdamamID() to Edamam service
First step of adjusting database to make it appropriate for custom recipes, edamam foods and favourites.

Initially, a call to edamam using the recipe ID to retrieve the recipe data was made in the storage.addRecipesFromStorage() however the code for this is now in the Edamam.getRecipeByEdamamID() service. This method is also now called by the storage.addRecipeToDay as this is the lowest level of the process.

The auth service has slightly been adjusted, the storage.addRecipeToDay call now includes passing the recipeID & recipeType  due to the change in logic within that method. Await has also been added in front of these calls as they're now asynchronous.

A few minor adjusts include the removal of console logging in ingredientDivider