Adjustments to Modal Boxes on Timetable Component, adjustment to account.service nutrients and start on new landing page
Modal Boxes on Timetable component

Each part of the main modal box now has it's own step as opposed to having a variable which is set to either to show buttons or content.

Additional steps have been put in place which allows the user to select which day they want to add the new meal to then which meal type they want to add (breakfast, snack etc.)

The calculateMacros() method has now been adjusted so that users will be recommended more fat with less protein and carbs.

Landing Page

A start on the new landing page has been made, this commit should show us how it looks on the mobile now.